ref: move generate hooks to openai router
